Updated project to support DHT11 and LED on off
Created a js class reptileData.js to store data in svr.js as a singleton.
EXPRESS Server updated for sending DHT11 to the client, control for Led Lights for the system also implemented.
Implemented get requests for Sending data to clients. Implemented a get request to run the lightSwitch function on the PI.
UI CSS edited, Snake logo Created.