code restructing and comment
Code has been fully commented, files have been restructured. 
Bash start script has been added, this can be called when the raspi is booting to automatically start the server.
An example of the generated log file has been committed. This is stored on the RasPi server and can later be used to allow the user to access a full history of all measured data regarding the reptile enclosure.